# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Storage and Send API\nFirst login with user and use authorization as \"Bearer JWTTOKEN\"",
        "contact": {}
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/login": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Get JWT token for 1 hour",
                "tags": [
                    "public"
                ],
                "summary": "Login with basic auth",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "raw token output",
                        "name": "raw",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apimodels.Data"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.LoginToken"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apimodels.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apimodels.Error"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/apimodels.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apimodels.Error"
                        }
                    }
                }
            },
            "post": {
                "description": "Get JWT token for 1 hour",
                "tags": [
                    "public"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "send login object",
                        "name": "payload",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/api.LoginModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apimodels.Data"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.LoginToken"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apimodels.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apimodels.Error"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/apimodels.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apimodels.Error"
                        }
                    }
                }
            }
        },
...
        "/run/js": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Run JS script with scripts and input values",
                "consumes": [
                    "text/plain"
                ],
                "tags": [
                    "run"
                ],
                "summary": "Run JS script",
                "parameters": [
                    {
                        "description": "Script and inputs, scriptB64 has priority over script",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/run.runModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "result of the script",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apimodels.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apimodels.Error"
                        }
                    }
                }
            }
        },
        "/run/template": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Render template with input values",
                "consumes": [
                    "text/plain"
                ],
                "tags": [
                    "run"
                ],
                "summary": "Render template",
                "parameters": [
                    {
                        "description": "send key values",
                        "name": "payload",
                        "in": "body",
                        "schema": {
                            "type": "string",
                            "example": ""
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apimodels.Data"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/apimodels.ID"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apimodels.Error"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/apimodels.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apimodels.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.LoginModel": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string",
                    "example": "admin"
                },
                "password": {
                    "type": "string",
                    "example": "admin"
                }
            }
        },
        "api.LoginToken": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string",
                    "example": "tokenJWT"
                }
            }
        },
...
        "apimodels.Data": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
...
        "apimodels.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "some problem"
                }
            }
        },
        "apimodels.ID": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "cf8a07d4-077e-402e-a46b-ac0ed50989ec"
                }
            }
        },
...
        "run.runModel": {
            "type": "object",
            "properties": {
                "inputs": {
                    "type": "array",
                    "items": {}
                },
                "script": {
                    "type": "string"
                },
                "scriptB64": {
                    "type": "string"
                },
                "settings": {
                    "$ref": "#/definitions/run.settings"
                }
            }
        },
        "run.settings": {
            "type": "object",
            "properties": {
                "timeout": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        }
    }
}